Fix compatibility issue with ST4114+
Fixes #213

This commit applies the required patch of ST4114's HTML.sublime-syntax
in order to fix VUE's incompatibility with most recent ST, while
maintaining backward compatibility with older builds.

The relevant contexts are overridden and variables are copied over
from HTML.sublime-syntax.

Note: The `style-close-tag` was probably not needed, as it hasn't
      changed, but it is copied to maintain formal consistency with
      script tags.