Relax test cases to avoid failure with ST <4114
The relaxed test cases would fail on <ST4114 as the related fix to avoid
duplicated meta scopes was introduced recently.

It's basically the same fix as for `lang-attribute` in previous commit,
but I don't think we should try to fix all kinds of issues of older
HTML.sublime-syntax.

This commit ensures to pass CI syntax tests.