Fix mustage template interpolation
This commit fixes implementation of mustage template tags {{...}}.

Before this commit they tried to pop the `main` context off stack
and caused `template-tag-content` to never have been popped off stack
properly.

With this commit the syntax distinguishes between embedded template
tags and those which interpolate strings. The latter one remove
`string` scope for better highlighting java code in `{{ ... }}`.

It also does no longer push a dedicated `mustage-content` context on
stack as this is the same content as the whole document. It's not
required.